requirements.txt
djangorestframework
django-cors-headers
djangorestframework-simplejwt
djongo==1.3.6
pymongo==3.12.3
pytz
dnspython
requests
groq
